pyproject.toml
[project]
name = "github-issue-log"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "pygithub>=2.8.1",
    "python-dateutil>=2.9.0.post0",
]
